[Add] Capture style.ts
